+++ src/modelgen/manager/IDataManager.java
@@ -7,5 +7,5 @@ import modelgen.data.property.PropertySettable;
 import modelgen.processor.IDataProcessorFactory;
 
 public interface IDataManager<I, O> extends PropertySettable {
-    Map.Entry<O, Integer> processData(I inputData, IDataProcessorFactory<I, O> processorFactory);
+    Map.Entry<O, Double> processData(I inputData, IDataProcessorFactory<I, O> processorFactory);
 }
